\name{SemiPar-internal}
\title{Internal SemiPar Objects}
\alias{add.model.flag}        
\alias{arg.search}            
\alias{blend.col}             
\alias{break.string}          
\alias{compute.CTRinvC}       
\alias{default.bdry}
\alias{default.knots} 
\alias{default.knots.2D} 
\alias{df.to.spar} 
\alias{dinv.link} 
\alias{dup}
\alias{dup.matrix} 
\alias{gen.corr.range} 
\alias{get.pixel.info} 
\alias{glm.df.to.spar}
\alias{glmeAux} 
\alias{imageFull} 
\alias{imageLegend} 
\alias{inv.link} 
\alias{irls.ridge} 
\alias{link}
\alias{lmeAux} 
\alias{lmeFitQr} 
\alias{matrix.sqrt} 
\alias{pixel.frame} 
\alias{plotControl}
\alias{pointsInPoly}  
\alias{random.read} 
\alias{ridge.reg}
\alias{rm.char} 
\alias{set.plot.dflts} 
\alias{spmArgRead} 
\alias{spmDesign} 
\alias{spmFormRead} 
\alias{spmKrigeRead} 
\alias{spmPenRead} 
\alias{spmTermType}
\alias{tps.cov} 
\description{Internal SemiPar functions.}
\details{These are not to be called by the user.}
\keyword{internal}



